refactor : Station 패키지 분리
