+++ src/main/java/subway/domain/station/Station.java
@@ -1,4 +1,4 @@
-package subway;
+package subway.domain.station;
 
 import javax.persistence.*;
