+++ src/main/java/subway/infrastructure/station/StationRepository.java
@@ -1,6 +1,7 @@
-package subway;
+package subway.infrastructure.station;
 
 import org.springframework.data.jpa.repository.JpaRepository;
+import subway.domain.station.Station;
 
 public interface StationRepository extends JpaRepository<Station, Long> {
 }
